+++ lib/src/sadja_progress_stepper_widget.dart
@@ -1,81 +1,184 @@
 import 'package:flutter/material.dart';
-import 'package:sadja_progress_stepper/src/step_item.dart';
 
-/// A customizable stepper widget that supports linear and non-linear navigation.
+import 'step_item.dart';
+
+/// A customizable stepper widget for Flutter that visually represents progress
+/// through multiple steps. Supports dynamic colors, icons, text, labels,
+/// and linear or non-linear navigation.
 ///
-/// This widget displays steps in a horizontal layout, allowing users to navigate
-/// through different steps by tapping on them. It supports:
-/// - **Linear Mode**: Steps must be completed in order.
-/// - **Non-Linear Mode**: Users can jump to any step freely.
+/// ## Features:
+/// - Customizable step indicators with colors, icons, and text.
+/// - Supports both **linear** (sequential) and **non-linear** (free) navigation.
+/// - Allows horizontal scrolling when steps exceed visible limits.
+/// - Optional step labels with static or dynamic color modes.
+/// - Configurable colors for steps, icons, text, and connecting lines.
 ///
-/// Example:
+/// ## Example Usage:
 /// ```dart
-/// SadjaProgressStepper(
-///   steps: [
-///     StepItem(icon: Icon(Icons.looks_one), content: Text("Step 1")),
-///     StepItem(icon: Icon(Icons.looks_two), content: Text("Step 2")),
-///   ],
-///   activeStepColor: Colors.blue,
-///   completedStepColor: Colors.green,
-///   incompleteStepColor: Colors.grey,
-///   completedSteps: [0],
-///   onStepTapped: (step) => print("Step $step tapped"),
+/// Center(
+///   child: SadjaProgressStepper(
+///     key: ValueKey(
+///       "$_currentStep $_completedSteps",
+///       ), // Forces a rebuild when _currentStep changes
+///     steps: steps!,
+///     currentStep: _currentStep,
+///     completedSteps: _completedSteps,
+///     activeStepColor: Colors.blue,
+///     completedStepColor: Colors.green,
+///     incompleteStepColor: Colors.grey,
+///     activeIconColor: Colors.white,
+///     completedIconColor: Colors.white,
+///     incompleteIconColor: Colors.black,
+///     activeTextColor: Colors.white,
+///     completedTextColor: Colors.white,
+///     incompleteTextColor: Colors.black,
+///     onStepTapped: _onStepTapped,
+///     showLabels: true,
+///   ),
 /// )
 /// ```
+///
+/// This widget is useful for onboarding, progress tracking, multi-step forms,
+/// and any process requiring step-by-step navigation.
 class SadjaProgressStepper extends StatefulWidget {
-  /// List of steps to be displayed in the stepper.
+  /// List of steps in the stepper.
   final List<StepItem> steps;
 
-  /// The index of the currently active step.
+  /// Index of the currently active step.
   final int currentStep;
 
   /// List of completed step indices.
   final List<int> completedSteps;
 
-  /// Color of the active step.
-  final Color activeStepColor;
-
-  /// Color of the completed steps.
-  final Color completedStepColor;
-
-  /// Color of the incomplete steps.
-  final Color incompleteStepColor;
-
   /// Callback function triggered when a step is tapped.
-  /// It provides the index of the selected step.
   final Function(int) onStepTapped;
 
-  /// Determines whether navigation should be **linear** (restricting access to future steps)
-  /// or **non-linear** (allowing free navigation).
+  /// Determines whether navigation is linear (restricts skipping steps).
   final bool isLinear;
 
-  /// Creates a `SadjaProgressStepper`.
+  /// Number of visible steps before scrolling is required.
+  final int visibleSteps;
+
+  // --- Step Indicator Customization ---
+
+  /// Background color for the active step.
+  final Color activeStepColor;
+
+  /// Background color for completed steps.
+  final Color completedStepColor;
+
+  /// Background color for incomplete steps.
+  final Color incompleteStepColor;
+
+  // --- Icon & Text Colors ---
+
+  /// Color of the icon in the active step.
+  final Color activeIconColor;
+
+  /// Color of the icon in completed steps.
+  final Color completedIconColor;
+
+  /// Color of the icon in incomplete steps.
+  final Color incompleteIconColor;
+
+  /// Text color for the active step.
+  final Color activeTextColor;
+
+  /// Text color for completed steps.
+  final Color completedTextColor;
+
+  /// Text color for incomplete steps.
+  final Color incompleteTextColor;
+
+  // --- Line Customization ---
+
+  /// Color of the line between the current step and the next step.
+  final Color? activeLineColor;
+
+  /// Color of the line between completed steps.
+  final Color? completedLineColor;
+
+  /// Color of the line between incomplete steps.
+  final Color? incompleteLineColor;
+
+  /// Whether the active step’s line should match the completed step color.
+  final bool activeLineFollowsCompletedColor;
+
+  // --- Label Customization ---
+
+  /// Determines whether labels should be shown below steps.
+  final bool showLabels;
+
+  /// Determines how label colors behave.
+  final LabelColorMode labelColorMode;
+
+  /// Static color for labels when `labelColorMode` is set to `LabelColorMode.static`.
+  final Color labelStaticColor;
+
+  /// Constructor for `SadjaProgressStepper`.
   ///
-  /// - `steps`: A list of `StepItem` widgets.
-  /// - `currentStep`: The active step index (default is 0).
-  /// - `completedSteps`: List of completed step indices.
-  /// - `activeStepColor`: The color of the currently active step.
-  /// - `completedStepColor`: The color of completed steps.
-  /// - `incompleteStepColor`: The color of incomplete steps.
-  /// - `onStepTapped`: Callback triggered when a step is tapped.
-  /// - `isLinear`: If `true`, users must complete steps in order.
+  /// The `steps` and `onStepTapped` are required, while `currentStep` defaults to `0`.
+  ///
+  /// Example:
+  /// ```dart
+  /// SadjaProgressStepper(
+  ///   steps: [...],
+  ///   currentStep: 2,
+  ///   completedSteps: [0, 1],
+  ///   onStepTapped: (index) => print(index),
+  ///   activeStepColor: Colors.blue,
+  ///   completedStepColor: Colors.green,
+  ///   incompleteStepColor: Colors.grey,
+  ///   activeIconColor: Colors.white,
+  ///   completedIconColor: Colors.white,
+  ///   incompleteIconColor: Colors.black,
+  ///   activeTextColor: Colors.white,
+  ///   completedTextColor: Colors.white,
+  ///   incompleteTextColor: Colors.black,
+  /// )
+  /// ```
   const SadjaProgressStepper({
     super.key,
     required this.steps,
     this.currentStep = 0,
     required this.completedSteps,
+    required this.onStepTapped,
+    this.isLinear = true,
+    this.visibleSteps = 3,
+
+    // Step Indicator Colors
     required this.activeStepColor,
     required this.completedStepColor,
     required this.incompleteStepColor,
-    required this.onStepTapped,
-    this.isLinear = true,
+
+    // Icon & Text Colors
+    required this.activeIconColor,
+    required this.completedIconColor,
+    required this.incompleteIconColor,
+    required this.activeTextColor,
+    required this.completedTextColor,
+    required this.incompleteTextColor,
+
+    // Line Colors
+    this.activeLineColor,
+    this.completedLineColor,
+    this.incompleteLineColor,
+    this.activeLineFollowsCompletedColor = false,
+
+    // Label Customization
+    this.showLabels = false,
+    this.labelColorMode = LabelColorMode.static,
+    this.labelStaticColor = Colors.grey,
   });
 
   @override
-  SadjaProgressStepperState createState() => SadjaProgressStepperState();
+  State<SadjaProgressStepper> createState() => _SadjaProgressStepperState();
 }
 
-class SadjaProgressStepperState extends State<SadjaProgressStepper> {
+/// Enum for controlling label color behavior
+enum LabelColorMode { static, dynamic }
+
+class _SadjaProgressStepperState extends State<SadjaProgressStepper> {
   late int _currentStep;
 
   @override
@@ -84,9 +187,13 @@ class SadjaProgressStepperState extends State<SadjaProgressStepper> {
     _currentStep = widget.currentStep;
   }
 
-  /// Determines the color of a step based on its status.
+  /// Determines the background color of the step based on its state
+  /// Returns the background color of a step based on its state.
+  /// - Active step: Uses `activeStepColor`
+  /// - Completed steps: Uses `completedStepColor`
+  /// - Incomplete steps: Uses `incompleteStepColor`
   Color _getStepColor(int index) {
-    if (index == widget.currentStep) {
+    if (index == _currentStep) {
       return widget.activeStepColor;
     } else if (widget.completedSteps.contains(index)) {
       return widget.completedStepColor;
@@ -95,115 +202,194 @@ class SadjaProgressStepperState extends State<SadjaProgressStepper> {
     }
   }
 
-  /// Handles navigation between steps.
-  ///
-  /// If `isLinear` is enabled, users cannot skip to an incomplete step.
-  void _goToStep(int step) {
-    // ✅ Restrict navigation if `isLinear` is enabled
-    if (widget.isLinear &&
-        step > 0 &&
-        !widget.completedSteps.contains(step - 1)) {
-      return;
+  /// Determines the icon color based on step state
+  /// Returns the icon color inside a step based on its state.
+  /// - Active step: Uses `activeIconColor`
+  /// - Completed steps: Uses `completedIconColor`
+  /// - Incomplete steps: Uses `incompleteIconColor`
+  Color _getIconColor(int index) {
+    if (index == _currentStep) {
+      return widget.activeIconColor;
+    } else if (widget.completedSteps.contains(index)) {
+      return widget.completedIconColor;
+    } else {
+      return widget.incompleteIconColor;
     }
-
-    setState(() {
-      _currentStep = step;
-    });
-    widget.onStepTapped(step); // ✅ Notify parent
   }
 
-  /// Builds an individual step indicator (circle with an icon).
+  /// Determines the text color based on step state
+  /// Returns the text color inside a step based on its state.
+  /// - Active step: Uses `activeTextColor`
+  /// - Completed steps: Uses `completedTextColor`
+  /// - Incomplete steps: Uses `incompleteTextColor`
+  Color _getTextColor(int index) {
+    if (index == _currentStep) {
+      return widget.activeTextColor;
+    } else if (widget.completedSteps.contains(index)) {
+      return widget.completedTextColor;
+    } else {
+      return widget.incompleteTextColor;
+    }
+  }
+
+  /// Determines the line color between steps
+  /// - If one of the steps is the current step, uses `activeLineColor` (or defaults to `completedStepColor`).
+  /// - If both steps are completed, uses `completedLineColor` (or defaults to `completedStepColor`).
+  /// - Otherwise, uses `incompleteLineColor` (or defaults to `incompleteStepColor`).
+  Color _getLineColor(int prevStep, int nextStep) {
+    bool isPrevCompleted = widget.completedSteps.contains(prevStep);
+    bool isNextCompleted = widget.completedSteps.contains(nextStep);
+
+    if (prevStep == _currentStep || nextStep == _currentStep) {
+      return widget.activeLineColor ?? widget.completedStepColor;
+    } else if (isPrevCompleted && isNextCompleted) {
+      return widget.completedLineColor ?? widget.completedStepColor;
+    } else {
+      return widget.incompleteLineColor ?? widget.incompleteStepColor;
+    }
+  }
+
+  /// Builds a single step indicator, which consists of:
+  /// - A circular step container that changes color based on step status.
+  /// - An icon or text inside the step.
+  /// - An optional label displayed below the step (if `showLabels` is enabled).
   Widget _buildStepIndicator(int index) {
-    return GestureDetector(
-      onTap: () => _goToStep(index),
-      child: Container(
-        width: 40,
-        height: 40,
-        decoration: BoxDecoration(
-          shape: BoxShape.circle,
-          color: _getStepColor(index),
+    final StepItem step = widget.steps[index];
+
+    return Column(
+      mainAxisAlignment: MainAxisAlignment.center,
+      crossAxisAlignment: CrossAxisAlignment.center,
+      children: [
+        GestureDetector(
+          onTap: () => _goToStep(index),
+          child: Container(
+            width: 50,
+            height: 50,
+            decoration: BoxDecoration(
+              shape: BoxShape.circle,
+              color: _getStepColor(index),
+            ),
+            child: Center(
+              child: step.text != null && step.text!.trim().isNotEmpty
+                  ? Text(
+                      step.text!.trim(),
+                      style: TextStyle(
+                        color: _getTextColor(index),
+                        fontWeight: FontWeight.bold,
+                      ),
+                    )
+                  : Icon(
+                      step.icon, // Always non-null
+                      color: _getIconColor(index),
+                      size: 30,
+                    ),
+            ),
+          ),
         ),
-        child: Center(child: widget.steps[index].icon),
-      ),
+
+        // Show label if enabled
+        if (widget.showLabels && step.label != null) ...[
+          const SizedBox(height: 4),
+          Text(
+            step.label!.trim(),
+            style: TextStyle(
+              color: widget.labelColorMode == LabelColorMode.static
+                  ? widget.labelStaticColor
+                  : _getStepColor(index),
+              fontSize: 12,
+            ),
+          ),
+        ],
+      ],
     );
   }
 
-  /// Builds the horizontal stepper layout.
-  /// with scrollable row in case of overflow
+  /// Builds the stepper layout with horizontal scrolling when necessary.
+  /// - Each step is displayed in sequence with a connecting line in between.
+  /// - Uses `SingleChildScrollView` to allow scrolling if the steps exceed the visible limit.
+  /// - Calculates dynamic spacing for steps and lines based on available width.
   Widget _buildScrollableStepper() {
+    debugPrint(
+        "Scroll Min required width ${(50 * widget.visibleSteps) + (100 * (widget.visibleSteps - 1))}");
     return LayoutBuilder(
       builder: (context, constraints) {
-        double screenWidth = constraints.maxWidth;
-        int minVisibleSteps = 4; // Ensure at least 4 steps are always visible
-        double stepWidth = 50; // Width of each step including padding
-        double lineWidth = 40; // Space between steps
+        double stepperWidth =
+            constraints.maxWidth * 0.9; // 90% of available width
+        double stepWidth = 60; // Width of each step
+        int numVisibleSteps = widget.visibleSteps;
 
-        double minRequiredWidth =
-            (stepWidth * minVisibleSteps) + (lineWidth * (minVisibleSteps - 1));
+        // Calculate dynamic line width based on available space
+        double totalStepWidth = stepWidth * numVisibleSteps;
+        double remainingWidth = stepperWidth - totalStepWidth;
+        double lineWidth = remainingWidth / (numVisibleSteps - 1);
 
-        bool needsScrolling = (widget.steps.length > minVisibleSteps) &&
-            (minRequiredWidth > screenWidth);
-
-        return SingleChildScrollView(
-          scrollDirection: Axis.horizontal,
-          child: Row(
-            mainAxisAlignment: MainAxisAlignment.start,
-            children: List.generate(widget.steps.length * 2 - 1, (index) {
-              if (index.isEven) {
-                // Step Indicator
-                int stepIndex = index ~/ 2;
-                return SizedBox(
-                  width: stepWidth,
-                  child: _buildStepIndicator(stepIndex),
-                );
-              } else {
-                // Line between steps
-                int prevStepIndex = (index - 1) ~/ 2;
-                int nextStepIndex = prevStepIndex + 1;
-
-                bool isPrevCompleted =
-                    widget.completedSteps.contains(prevStepIndex);
-                bool isNextCompleted =
-                    widget.completedSteps.contains(nextStepIndex);
-
-                Color lineColor = (isPrevCompleted && isNextCompleted)
-                    ? widget.completedStepColor
-                    : widget.incompleteStepColor;
-
-                return SizedBox(
-                  width: needsScrolling
-                      ? lineWidth
-                      : (screenWidth - stepWidth * widget.steps.length) /
-                          (widget.steps.length - 1),
-                  height: 4,
-                  child: DecoratedBox(
-                    decoration: BoxDecoration(color: lineColor),
-                  ),
-                );
-              }
-            }),
+        return SizedBox(
+          width: stepperWidth,
+          child: SingleChildScrollView(
+            scrollDirection: Axis.horizontal,
+            child: Row(
+              crossAxisAlignment: CrossAxisAlignment.center,
+              mainAxisAlignment: MainAxisAlignment.center,
+              children: List.generate(
+                widget.steps.length * 2 - 1,
+                (index) {
+                  if (index.isEven) {
+                    int stepIndex = index ~/ 2;
+                    return SizedBox(
+                      width: stepWidth,
+                      child: _buildStepIndicator(stepIndex),
+                    );
+                  } else {
+                    return Container(
+                      width: lineWidth > 0 ? lineWidth : 10,
+                      height: 4, // Keep line thickness
+                      color: _getLineColor(
+                          (index - 1) ~/ 2, (index + 1) ~/ 2), // Set line color
+                    );
+                  }
+                },
+              ),
+            ),
           ),
         );
       },
     );
   }
 
+  /// Handles navigation when a step is tapped.
+  /// - If `isLinear` is enabled, prevents skipping steps.
+  /// - Updates `_currentStep` and triggers `onStepTapped` callback.
+  void _goToStep(int step) {
+    // Restrict navigation if `isLinear` is enabled
+    if (widget.isLinear &&
+        step > 0 &&
+        !widget.completedSteps.contains(step - 1)) {
+      debugPrint("Step $step is locked! Complete previous steps first.");
+      return;
+    }
+
+    setState(() {
+      _currentStep = step;
+    });
+
+    // Notify parent about step change
+    widget.onStepTapped(step);
+  }
+
   @override
   Widget build(BuildContext context) {
-    return Container(
-      margin: EdgeInsets.symmetric(horizontal: 2, vertical: 2),
-      color: Colors.white10,
-      child: Column(
-        children: [
-          // ✅ Use scrollable stepper instead of a fixed row
-          _buildScrollableStepper(),
+    return Column(
+      children: [
+        // Stepper Header (Step Indicators + Connecting Lines)
+        _buildScrollableStepper(),
 
-          // ✅ Display the content of the currently active step
-          Expanded(
-            child: widget.steps[_currentStep].content,
-          ),
-        ],
-      ),
+        const SizedBox(height: 16),
+
+        // Step Content (Widget that appears when a step is selected)
+        Expanded(
+          child: widget.steps[_currentStep].content,
+        ),
+      ],
     );
   }
 }